renamed kernelsharedtable package and removed not used variables
