+++ server/src/main/scala/io/delta/standalone/internal/DeltaSharedTable.scala
@@ -20,7 +20,6 @@ package io.delta.standalone.internal
 import java.net.URI
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.Base64
-import java.util.concurrent.TimeUnit
 
 import scala.collection.JavaConverters._
 
@@ -41,7 +40,7 @@ import scala.collection.mutable.ListBuffer
 import scala.util.control.NonFatal
 import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
 
-import io.delta.sharing.server.{model, CausedBy, DeltaSharedTableProtocol, DeltaSharingIllegalArgumentException, DeltaSharingUnsupportedOperationException, ErrorStrings, QueryResult}
+import io.delta.sharing.server.{model, DeltaSharedTableProtocol, DeltaSharingIllegalArgumentException, DeltaSharingUnsupportedOperationException, ErrorStrings, QueryResult}
 import io.delta.sharing.server.common.{AbfsFileSigner, GCSFileSigner, PreSignedUrl, S3FileSigner, WasbFileSigner}
 import io.delta.sharing.server.config.TableConfig
 import io.delta.sharing.server.protocol.{QueryTablePageToken, RefreshToken}
